requirements.txt
aiogram==2.25.2
pymongo[srv]
openai
python-dotenv
